sort imports

+++ src/app/transformation.py
@@ -1,6 +1,6 @@
 import numpy as np
 
-from src.logger import logger
+from ..logger import logger
 
 
 class Transformation:
@@ -9,10 +9,9 @@ class Transformation:
         logger.log(f"upscaling array by factor {scalar}")
         big = np.zeros(
             (arr.shape[0] * scalar, arr.shape[1] * scalar, arr.shape[2])
-            # dtype=arr.dtype,
         )
         big[0::scalar, 0::scalar] = arr
         big[1::scalar, 1::scalar] = arr
         big[0::scalar, 1::scalar] = arr
         big[1::scalar, 0::scalar] = arr
-        return big
+        return big.astype(np.uint8)
